refactor(#12408): move view-interact protocol constants into @elizaos/shared
AGENT_SURFACE_CAPABILITY_IDS and STANDARD_CAPABILITIES were defined inside
@elizaos/ui internals, but the agent views route imported them via undeclared,
hoisting-only @elizaos/ui specifiers — coupling the standalone-publishable agent
server to UI internals it does not depend on.

- Add packages/shared/src/views/view-interact-protocol.ts as the single
  canonical source for the view-interact contract (STANDARD_CAPABILITIES,
  AGENT_SURFACE_CAPABILITY_IDS, ViewInteractRequest/Result, StandardCapability),
  exported via the "@elizaos/shared/views/view-interact-protocol" subpath and the
  shared barrel.
- @elizaos/ui now re-exports the constants from shared, so the historical
  @elizaos/ui import paths keep working for UI consumers with zero drift.
- packages/agent/src/api/views-routes.ts imports both constants from
  @elizaos/shared; no more @elizaos/ui import statements in packages/agent/src.
- Add packages/shared/src/views/view-interact-protocol.test.ts pinning the
  frozen capability ids/values and their disjointness.

Closes #12408
