+++ packages/ui/src/views/view-interact-protocol.ts
@@ -1,49 +1,14 @@
 /**
- * View interact protocol — shared types for agent ↔ view interactions.
- *
- * The flow:
- *   1. Agent POSTs to /api/views/:id/interact with a ViewInteractRequest body.
- *   2. Server broadcasts a WS message {type:"view:interact", ...} to all clients.
- *   3. DynamicViewLoader receives the WS message, calls the view module's
- *      interact(capability, params) export (or a standard capability handler).
- *   4. Frontend sends {type:"view:interact:result", ...} back over WS.
- *   5. Server resolves the pending request and returns the result to the agent.
+ * Re-exports the canonical view-interact protocol contract from
+ * `@elizaos/shared`. The constants and types live in shared so the agent server
+ * can dispatch against them without importing UI internals (#12408); this module
+ * keeps the historical `@elizaos/ui/views/view-interact-protocol` import path
+ * working for UI consumers.
  */
 
-export interface ViewInteractRequest {
-  viewId: string;
-  capability: string;
-  params?: Record<string, unknown>;
-  /** UUID generated server-side for correlating the async result. */
-  requestId: string;
-  /** Timeout in ms before the server gives up waiting. Default 5000. */
-  timeoutMs?: number;
-}
-
-export interface ViewInteractResult {
-  requestId: string;
-  success: boolean;
-  result?: unknown;
-  error?: string;
-}
-
-/** Standard capabilities that every view is expected to support. */
-export const STANDARD_CAPABILITIES = {
-  /** Returns the current view state as JSON. */
-  GET_STATE: "get-state",
-  /** Forces a data refresh / re-render. */
-  REFRESH: "refresh",
-  /** Focuses an input or button by CSS selector or name attribute. */
-  FOCUS_ELEMENT: "focus-element",
-  /** Returns the visible text content of the view container. */
-  GET_TEXT: "get-text",
-  /** Clicks an element by CSS selector or name attribute. Dispatched generically
-   *  by DynamicViewLoader / ShellViewAgentSurface for every loaded view. */
-  CLICK_ELEMENT: "click-element",
-  /** Sets the value of an input by selector/name. Dispatched generically by
-   *  DynamicViewLoader / ShellViewAgentSurface for every loaded view. */
-  FILL_INPUT: "fill-input",
-} as const;
-
-export type StandardCapability =
-  (typeof STANDARD_CAPABILITIES)[keyof typeof STANDARD_CAPABILITIES];
+export {
+  STANDARD_CAPABILITIES,
+  type StandardCapability,
+  type ViewInteractRequest,
+  type ViewInteractResult,
+} from "@elizaos/shared/views/view-interact-protocol";
